+++ lib/background/game_background.dart
@@ -5,4 +5,10 @@ import 'package:bonfire/util/priority_layer.dart';
 class GameBackground extends GameComponent {
   @override
   int get priority => LayerPriority.BACKGROUND;
+
+  @override
+  bool get enabledCheckIsVisible => false;
+
+  @override
+  bool get isVisible => true;
 }