+++ lib/base/bonfire_game.dart
@@ -6,6 +6,7 @@ import 'package:bonfire/base/base_game.dart';
 import 'package:bonfire/bonfire.dart';
 import 'package:bonfire/camera/bonfire_camera.dart';
 import 'package:bonfire/color_filter/color_filter_component.dart';
+import 'package:bonfire/input/keyboard/control_by_keyboard.dart';
 import 'package:bonfire/joystick/joystick_map_explorer.dart';
 import 'package:bonfire/lighting/lighting_component.dart';
 // ignore: implementation_imports
@@ -35,7 +36,7 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
   final GameMap map;
 
   /// The player-controlling component.
-  final JoystickController? joystickController;
+  final PlayerController? joystickController;
 
   /// Background of the game. This can be a color or custom component
   final GameBackground? background;
@@ -53,6 +54,8 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
 
   @override
   final List<Force2D> globalForces;
+
+  final KeyboardConfig? keyboardConfig;
 
   @override
   SceneBuilderStatus sceneBuilderStatus = SceneBuilderStatus();
@@ -73,7 +76,7 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
       camera.viewport.children.whereType<ColorFilterInterface>().first;
 
   @override
-  JoystickController? get joystick => joystickController;
+  PlayerController? get joystick => joystickController;
 
   @override
   Color backgroundColor() => _bgColor ?? super.backgroundColor();
@@ -101,6 +104,7 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
     required this.context,
     required this.map,
     this.joystickController,
+    this.keyboardConfig,
     this.player,
     this.interface,
     List<GameComponent>? components,
@@ -121,6 +125,7 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
             viewport: cameraConfig?.resolution != null
                 ? FixedResolutionViewport(resolution: cameraConfig!.resolution!)
                 : null,
+            backdrop: background,
             hudComponents: [
               LightingComponent(
                 color: lightingColorGame ?? const Color(0x00000000),
@@ -129,13 +134,15 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
                 colorFilter ?? GameColorFilter(),
               ),
               if (joystickController != null) joystickController,
+              ControlByKeyboard(
+                keyboardConfig: keyboardConfig,
+              ),
               if (interface != null) interface,
             ],
           ),
           world: World(
             children: [
               map,
-              if (background != null) background,
               if (player != null) player,
               ...components ?? [],
             ],
@@ -161,9 +168,11 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
       mapDimensions: Rect.zero,
     );
 
-    joystickController?.addObserver(
-      player ?? JoystickMapExplorer(camera),
-    );
+    camera.viewport.children.query<PlayerController>().forEach((element) {
+      element.addObserver(
+        player ?? JoystickMapExplorer(camera),
+      );
+    });
 
     if (camera.config.target != null) {
       camera.follow(
@@ -196,8 +205,7 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
   @override
   void onMount() {
     super.onMount();
-    // ignore: invalid_use_of_internal_member
-    setMounted();
+    _notifyGameMounted();
     onReady?.call(this);
   }
 
@@ -231,12 +239,9 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
   Iterable<ShapeHitbox> collisions({bool onlyVisible = false}) {
     if (onlyVisible) {
       List<ShapeHitbox> tilesCollision = [];
-      map
-          .getRendered()
-          .where((element) => element.containsShapeHitbox)
-          .forEach((e) {
-        tilesCollision.addAll(e.children.query<ShapeHitbox>());
-      });
+      map.getRendered().where((element) => element.containsShapeHitbox).forEach(
+            (e) => tilesCollision.addAll(e.children.query<ShapeHitbox>()),
+          );
       return [
         ..._visibleCollisions,
         ...tilesCollision,
@@ -273,7 +278,7 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
   /// Use this method to change default observer of the Joystick events.
   @override
   void addJoystickObserver(
-    JoystickListener target, {
+    PlayerControllerListener target, {
     bool cleanObservers = false,
     bool moveCameraToTarget = false,
   }) {
@@ -306,13 +311,10 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
 
   @override
   void onDetach() {
-    world.children.query<GameComponent>().forEach(_detachComp);
-    camera.viewport.children.query<GameComponent>().forEach(_detachComp);
+    _notifyGameDetach();
     FollowerWidget.removeAll();
     super.onDetach();
   }
-
-  void _detachComp(GameComponent c) => c.onGameDetach();
 
   void addVisible(GameComponent obj) {
     _visibleComponents.add(obj);
@@ -458,4 +460,29 @@ class BonfireGame extends BaseGame implements BonfireGameInterface {
       scheduleMicrotask(_updateOrderPriority);
     }
   }
+
+  void _notifyGameMounted() {
+    void gameMontedComp(GameComponent c) => c.onGameMounted();
+    query<GameComponent>().forEach(gameMontedComp);
+    for (var child in camera.children) {
+      if (child is GameComponent) {
+        child.onGameMounted();
+      }
+      child.children.query<GameComponent>().forEach(gameMontedComp);
+    }
+  }
+
+  void _notifyGameDetach() {
+    void gameDetachComp(GameComponent c) => c.onGameDetach();
+    query<GameComponent>().forEach(gameDetachComp);
+    for (var child in camera.children) {
+      if (child is GameComponent) {
+        child.onGameDetach();
+      }
+      child.children.query<GameComponent>().forEach(gameDetachComp);
+    }
+  }
+
+  @override
+  Vector2 get worldsize => map.size;
 }